+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 (Add entries below in the order of server, console, cli, docs, others)
 
 - server: Fixing bug in ReplaceMetadata parser - Moving from Alternative to committed-choice.
+- server: Relax the unique operation name constraint when adding a query to a query collection
 
 ## v2.0.3
 (Add entries below in the order of server, console, cli, docs, others)